Designing module for additional functionality
