+++ QvPattern/MainWindow.xaml
@@ -1,0 +1,15 @@
+<main:ModuleWindow x:Class="QviKD.Module.MainWindow"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:QviKD.Module"
+             xmlns:main="clr-namespace:QviKD;assembly=QviKD"
+             mc:Ignorable="d" 
+             d:DesignHeight="450" d:DesignWidth="800"
+             Title="QvPattern">
+    <StackPanel>
+        <TextBlock>Hello World!</TextBlock>
+        <Button>Click</Button>
+    </StackPanel>
+</main:ModuleWindow>
