+++ QviKD/Function/EnumModules.cs
@@ -1,10 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Reflection;
 using System.IO;
+using System.Diagnostics;
 
 namespace QviKD.Functions
 {
@@ -12,25 +10,41 @@ namespace QviKD.Functions
     {
         public EnumModules(string directory = null)
         {
+            Database.Modules.Clear();
+            
             // Set QviKD.exe directory by default.
             if (directory == null)
             {
                 directory = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
             }
+            DebugMessage($"Module directory: {directory}");
 
             // Acquire DLL assemblies excluding QviKD.dll.
-            List<string> modules = Directory.EnumerateFiles(
+            List<string> module = Directory.EnumerateFiles(
                 directory,
                 "*.dll",
                 SearchOption.TopDirectoryOnly).Where(path => path != Assembly.GetExecutingAssembly().Location
                 ).ToList();
 
             // Enumerate over the DLL assemblies.
-            for (int index = 0; index < modules.Count; index++)
+            for (int index = 0; index < module.Count; index++)
             {
-                Database.Modules.Add(Assembly.LoadFile(modules[index]));
-            }
+                Assembly assembly = Assembly.LoadFile(module[index]);
 
+                // Add valid module assembly to the List.
+                if (Module.IsValidModule(assembly))
+                {
+                    Database.Modules.Add(new Module(assembly));
+                    DebugMessage($"Module: {assembly.FullName}");
+                }
+            }
+        }
+
+        private void DebugMessage(string msg)
+        {
+#if DEBUG
+            Debug.WriteLine($"'{GetType().Name}.cs' {msg}");
+#endif
         }
 
     }
