+++ QviKD/ModulePage.xaml
@@ -1,4 +1,4 @@
-﻿<Page x:Class="QviKD.ModulesPage"
+﻿<Page x:Class="QviKD.ModulePage"
       xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
       xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
       xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
